requirements.txt
requests==2.32.5
werkzeug==3.1.3
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
bcrypt==5.0.0
cerberus==1.3.8
PyJWT==2.10.1